[Refactor] 사진 편집 / 추가 관련 상태를 보기 좋게 변경
- Bool 타입의 플래그 2개를 번갈아 사용하던 형태를, 
- enum 형태로 state 를 관리해서 명시적이게 변경함
