+++ EasyCloset/EasyCloset/Sources/View/AddPhotoButton.swift
@@ -6,7 +6,9 @@
 //
 
 import UIKit
+
 import SnapKit
+import Then
 
 final class AddPhotoButton: UIButton {
