--- EasyCloset/EasyCloset/Sources/View/AddPhotoView.swift
@@ -1,8 +1,0 @@
-//
-//  AddPhotoView.swift
-//  EasyCloset
-//
-//  Created by Mason Kim on 2023/05/21.
-//
-
-import Foundation
